requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-pptx==0.6.21
python-multipart==0.0.6
openai==1.3.0
anthropic==0.14.0
google-generativeai==0.3.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic-settings==2.1.0
requests==2.31.0
pillow==10.1.0
xmltodict==0.13.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv